+++ Webserv_sketch/Sketch2/FileDescriptor/IOnEvents.hpp
@@ -1,26 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   IPollableSocket.hpp                                :+:      :+:    :+:   */
+/*   IOnEvents.hpp                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mmaria-d <mmaria-d@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/09/16 14:57:22 by mmaria-d          #+#    #+#             */
-/*   Updated: 2024/09/16 15:03:41 by mmaria-d         ###   ########.fr       */
+/*   Created: 2024/09/16 14:55:53 by mmaria-d          #+#    #+#             */
+/*   Updated: 2024/09/16 15:34:00 by mmaria-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef IPOLLABLEFILEDESCRIPTOR_HPP
+#ifndef IONEVENTS_HPP
 
-# define IPOLLABLEFILEDESCRIPTOR_HPP
+# define IONEVENTS_HPP
 
-# include "../../FileDescriptor/FileDescriptor.hpp"
-# include "IOnEvents.hpp"
-
-class IPollableFileDescriptor : public FileDescriptor, public IOnEvents
+class IOnEvents
 {
     public:
-        virtual ~IPollableFileDescriptor() {}
+        virtual ~IOnEvents() {}
+
+        virtual void onClose() = 0;
+        virtual void onRead() = 0;
+        virtual void onWrite() = 0;
+        virtual void onError() = 0;
 };
