+++ Webserv_sketch/Sketch2/Socket/Abstract/ISocket.hpp
@@ -6,7 +6,7 @@
 /*   By: mmaria-d <mmaria-d@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/16 10:59:48 by mmaria-d          #+#    #+#             */
-/*   Updated: 2024/09/16 15:04:22 by mmaria-d         ###   ########.fr       */
+/*   Updated: 2024/09/16 15:30:21 by mmaria-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 # define ISOCKET_HPP
 
-# include "IPollableFileDescriptor.hpp"
+# include "../../FileDescriptor/FileDescriptor.hpp"
 # include "ISocketAddress.hpp"
 
-class ISocket : public IPollableFileDescriptor, public ISocketAddress
+class ISocket : public FileDescriptor, public ISocketAddress
 {
     public:
         virtual         ~ISocket() {}
